requirements_file.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
ultralytics>=8.0.0
numpy>=1.21.0
matplotlib>=3.3.0
networkx>=2.6.0
scipy>=1.7.0
Pillow>=8.3.0
pyyaml>=5.4.0
tqdm>=4.62.0
scikit-learn>=1.0.0
seaborn>=0.11.0
